Bunch of small changes/fixes (#264)
* Fixed EXCON time length being converted improperly

* Adjusted slide timings for reverb

* Added hard coded multiplier for song volume to help with some balancing

* Adjusted some default volume settings

* Added 2 second delay before song begins

* Added function to scan multiple folders (not all or just 1)

* Fixed preparsers not assigning parts beyond 32 bits correctly

* Updated HasPart to use long bits

+++ Assets/Script/Song/Preparsers/ChartPreparser.cs
@@ -70,7 +70,7 @@ namespace YARG.Song.Preparsers {
 				}
 					
 				int shiftAmount = (int)track.instrument * 4 + (int)track.difficulty;
-				tracks |= (uint)(1 << shiftAmount);
+				tracks |= 1UL << shiftAmount;
 			}
 
 			return tracks;
